% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wimtrap.R
\name{buildTFBSmodel}
\alias{buildTFBSmodel}
\title{Builds a predictive model of transcription factor binding site (TFBS) location}
\usage{
buildTFBSmodel(TFBSdata, TF4Validation = NULL, model_assessment = TRUE)
}
\arguments{
\item{TFBSdata}{A data.table object as output by the \code{\link[=getTFBSdata]{getTFBSdata()}} function.}

\item{TF4Validation}{NULL (by default) or character vector of names of training TFs. If NULL, the model performances will be
assessed from a validation dataset obtained by sampling randomly 20\% of the potential binding sites of all the training TFs.
Otherwise the model will be validated with the data related to the training TFs named in the \emph{TF4Validation} parameter and
trained using the remaining training TFs. Please note, that in the last case, the \emph{TFBSdata} table must contain a
column named 'TF'.}

\item{model_assessment}{Logical. If TRUE (by default), 1) the imortance of features is represented and, based on the
validation dataset, 2) the confusion matrix is printed and 3) the ROC and AUC are plotted.}
}
\value{
An object of xgb.Booster class.
}
\description{
\code{buildTFBSmodel} learns (and optionally validates) a predictive model of transcription factor(TF) binding sites based
on the integration of genomic features extracted at the location of potential binding
sites identified by a prior analysis (such as pattern-matching). This function implements
an algorithm of supervised machine learning, the extreme gradient boosting (XGBoost), which
is fed by a dataset of potential binding sites of training TFs either labelled as 'positive' or 'negative'
(i.e. ChIP-seq 'validated' or 'not-validated' in a given condition).
}
\seealso{
\code{\link[=getTFBSdata]{getTFBSdata()}} for obtaining a training dataset and \code{\link[=predictTFBS]{predictTFBS()}} to predict transcription factor
binding site location
}
