% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wimtrap.R
\name{plotPredictions}
\alias{plotPredictions}
\title{Plots genomic data and predicted binding sites along a gene track}
\usage{
plotPredictions(
  TFBSpredictions,
  imported_genomic_data,
  gene,
  TFs = NULL,
  genomic_data = NULL,
  promoter_length = 2000,
  downstream_length = 2000
)
}
\arguments{
\item{imported_genomic_data}{A list of \code{GRanges} objects as output by \code{\link[=importGenomicData]{importGenomicData()}}. It includes at least
the three \code{GRanges} named '\code{CDS}', '\code{X5UTR}' and '\code{X3UTR}' that give the position of the coding sequence(s), 5'UTR and
3'UTR of the transcript models that are indicated in the '\code{name}' metadata column.}

\item{gene}{A character. Sets the name of the gene to be plotted. The system of nomenclature corresponds to that
used to name the transcript models in the \code{GRanges} objects of \code{imported_genomic_data}.}

\item{TFs}{\code{NULL} or a vector of characters indicating the transcription factors for which the predicted binding
sites have to be plotted. By default, all the studied transcription factors will be considered.}

\item{genomic_data}{\code{NULL} or a vector of characters allowing to name the genomic data for which the signal has to
be plotted. The names that are accepted correspond to those of the GRanges objects listed in \code{imported_genomic_data}.
By default, no genomic data is plotted.}

\item{promoter_length}{A numeric allowing to set the start of the gene track (=lowest start among the transcript
models of the considered gene - \code{promoter_length}). Default = 2000.}

\item{downstream_length}{A numeric allowing to set the end of the gene track (=highest among the transcript
models of the considered gene - \code{downstreamr_length}). Default = 2000.}

\item{TFBSprediction}{A \code{data.table} as output by \code{\link[=predictTFBS]{predictTFBS()}}. This \code{data.table} comprises the columns '\code{seqnames}',
'\code{start}', '\code{end}', '\code{width}', '\code{strand}', '\code{transcript}', '\code{prediction.score}', '\code{TF}'
and, optionally, the annotations with the predictive features extracted from the genomic data.}
}
\value{
A list of \code{GenomeGraph} tracks.
}
\description{
Allows the vizualisation of the predicted binding sites on a given gene for the studied condition.
The function plots three panels of gene tracks: (i) the different transcript models of the gene of interest;
(ii) the predicted binding sites, annotated with their prediction score; and (iii), optionally, the signals
of the genomic data that have been used to extract the predictive features at the location of the potential
binding sites.
}
\examples{
genomic_data.ex <- c(CE = system.file("extdata/conserved_elements_example.bed", package = "Wimtrap"),
                      DGF = system.file("extdata/DGF_example.bed", package = "Wimtrap"),
                      DHS = system.file("extdata/DHS_example.bed", package = "Wimtrap"),
                      X5UTR = system.file("extdata/x5utr_example.bed", package = "Wimtrap"),
                      CDS = system.file("extdata/cds_example.bed", package = "Wimtrap"),
                      Intron = system.file("extdata/intron_example.bed", package = "Wimtrap"),
                      X3UTR = system.file("extdata/x3utr_example.bed", package = "Wimtrap")
                     )
imported_genomic_data.ex <- importGenomicData(biomart = FALSE,
                                              genomic_data = genomic_data.ex,
                                              tss = system.file("extdata/tss_example.bed", package = "Wimtrap"),
                                              tts = system.file("extdata/tts_example.bed", package = "Wimtrap"))
TFBSdata.ex <- getTFBSdata(pfm = system.file("extdata/pfm_example.pfm", package = "Wimtrap"),
                           TFnames = c("PIF3", "TOC1"),
                           organism = NULL,
                           genome_sequence = system.file("extdata/genome_example.fa", package = "Wimtrap"),
                           imported_genomic_data = imported_genomic_data.ex)
TFBSmodel.ex <- buildTFBSmodel(TFBSdata = TFBSdata.ex,
                               ChIPpeaks = c(PIF3 = system.file("extdata/PIF3_example.bed", package = "Wimtrap"),
                                             TOC1 = system.file("extdata/TOC1_example.bed", package = "Wimtrap")),
                               TFs_validation = "PIF3")
PIF3BS.predictions <- predictTFBS(TFBSmodel.ex,
                                  TFBSdata.ex,
                                  studiedTFs = "PIF3")
plotPredictions(PIF3BS.predictions,
                imported_genomic_data.ex,
                gene = "AT1G01140",
                genomic_data = "DHS")
}
\seealso{
\code{\link[=predictTFBS]{predictTFBS()}} to get predictions of transcription factor binding sites.
}
