% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wimtrap.R
\name{getTFBSdata}
\alias{getTFBSdata}
\title{Builds a dataset that allows to train a model of transcription factor binding sites (TFBS)}
\usage{
getTFBSdata(
  ChIPpeaks,
  pfm = NULL,
  organism = NULL,
  genome_sequence = "getGenome",
  imported_genomic_data,
  matches = NULL,
  strandAsFeature = FALSE,
  pvalThreshold = 0.001
)
}
\arguments{
\item{ChIPpeaks}{A named character vector defining the local paths to BED files encoding the
location of ChIP-peaks. The vector is named according to the transcription factors that are described
by the files indicated.}

\item{pfm}{Path to a file in the raw Jaspar format. This file describes the Pseudo Frequency Matrices (PFMs)
of the primary motifs of the transcription factors whose ChIP-seq data are considered.
IMPORTANT: In the file indicated, the PFMs have to be named by their cognate transcription
factors. These names have to be consistent with those provided through the \emph{ChIP-peaks} argument.
\emph{pfm} can be set to NULL if you prefer to use pattern-matching results from elsewhere (see the
argument \emph{matches}).}

\item{organism}{Binomial name of the organism. Can be set to NULL if you provide the genome sequence
(see the argument \emph{genome_sequence})}

\item{genome_sequence}{"getGenome" (by default) or local path to a FASTA file encoding the genomic sequence
of the organism. The default value allows the automatic download of the genomic sequence (when \emph{organism} is defined)
from ENSEMBL or ENSEMBL GENOMES.}

\item{imported_genomic_data}{An object output by \code{\link[=importGenomicData]{importGenomicData()}} and that includes data related to the chromatin
state that are specific to the condition in which the ChIP-seq data considered have been obtained.}

\item{matches}{NULL (by default) or a named list of GRanges objects. Each GRanges object defines the location along
the genome of the matches with the primary motif of a given transcription factor. It contains also a metadata column
named 'matchLogPval'that gives the p-value of the matches. The list input through \emph{matches} has to be named according
to the names of the transcription factors considered. These names have to be consistent with those provided through
the \emph{ChIP-peaks} argument. The default value allows to perform the pattern-matching analysis with the function encoded
by the Wimtrap package.}

\item{strandAsFeature}{Logical. Should be considered as feature the orientation of the matches in relation to the
direction of transcription of the closest transcript? Default is FALSE.}

\item{pvalThreshold}{P-value threshold to identify the matches with the primary motifs of the transcription factors.
Default is set to 0.001.}

\item{short_window}{Integer (20 by default). Sets the length of the short-ranges window centered on the potential binding
sites and on which the genomic features are extracted.}

\item{medium_window}{Integer (400 by default). Sets the length of the medium-ranges window centered on the potential binding
sites and on which the genomic features are extracted.}

\item{long_window}{Integer (1000 by default). Sets the length of the long-ranges window centered on the potential binding
sites and on which the genomic features are extracted.}
}
\value{
A data.table describing the location of the potential binding sites in the 5 first columns, the
results of pattern-matching (raw score and/or p-value), the genomic features extracted on short-, medium-
and long-ranges windows centered on the potential binding sites, the label ('1' = "positive" = "ChIP-validated in
a given condition" or '0' = "negative") and the name of the cognate transcription factor.
}
\description{
\code{getTFBSdata} outputs a balanced dataset characterizing the genomic context
around occurences of primary motifs that have been proved to be either TF-bound
or TF-not bound sites according to ChIP-seq data obtained in given experimental conditions.
}
\seealso{
\code{\link[=importGenomicData]{importGenomicData()}} for importing genomic data and \code{\link[=buildTFBSmodel]{buildTFBSmodel()}} to train a predictive model of
transcription factor binding sites.
}
