% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wimtrap.R
\name{importGenomicData}
\alias{importGenomicData}
\title{Imports genomic data into the R session}
\usage{
importGenomicData(
  organism = NULL,
  genomic_data,
  biomart = TRUE,
  tss = NULL,
  tts = NULL,
  promoterLength = 2000,
  downstreamLength = 1000,
  proximalLength = 500
)
}
\arguments{
\item{organism}{Binomial name of the organism. Can be set to NULL if you provide
the location of the transcription start sites (TSS), transcription termination site (TTS) and structures
of the protein-coding genes of the organism (see the arguments  \emph{tss}, \emph{tts} and \emph{genomic_data})}

\item{genomic_data}{A named character vector defining the local paths to BED files describing genomic features.
The vector has to be named according to the features described by the files indicated. All the data related to
the chromatin state have to be specific of the samed condition. The proporties of the BED files are the following,
depending on the type of feature: 'numeric' (the score field of the file is fulfilled or empty - if empty, the
score will automatically be set to '1') or 'categorical' (the score field of the file is empty while its name field
is fulfilled with the names of the categories) features.}

\item{biomart}{Logical. Should be automatically downloaded through biomart the location of the transcription
start sites (TSS), transcription termination site (TTS) and structures of the protein-coding genes of the organism?
Default is TRUE.}

\item{tss}{NULL (by default) or local path to a BED file defining the transcription stat site (TSS), name and
orientation of each protein-coding transcript of the organism. The default value allows to download automatically
these informations from biomart.}

\item{tts}{NULL (by default) or local path to a BED file defining the transcription termination site (TTS), name
and orientation of each protein-coding transcript of the organism. The default value allows to download automatically
these informations from biomart.}

\item{promoterLength}{Length of the promoters. By default, the promoter is defined as the region spanning the 2000bp
upstream of the transcription start site (TSS).}

\item{downstreamLength}{Length of the downstream regions. By default, the downstream region is defined as the region
spanning the 1000bp downstream of the transcription termnination site (TTS).}

\item{proximalLength}{Length of the proximal promoters. By default, the proximal promoter is defined as the region
spanning the 500bp upstream of the transcription start site (TSS).}
}
\value{
A named list of GRanges objects. Each component of the list is named according to the nature of the genomic
data. The last two components describe respectively the position of the transcritpion termination site (TTS) and the
transcription start site (TSS) of each protein-coding transcripts of the organism considered.
}
\description{
Imports genomic data that will allow to define contextual features around matches with the primary motifs of
transcrription factors. Data about gene structures might be optionally automatically downloaded from Biomart.
}
\examples{
## Without automatic download from Biomart of data related to gene structure
genomic_data.ex <- c(CE = system.file("conserved_elements_example.bed", package = "Wimtrap"),
                      DGF = system.file("DGF_example.bed", package = "Wimtrap"),
                      DHS = system.file("DHS_example.bed", package = "Wimtrap"),
                      X5UTR = system.file("x5utr_example.bed", package = "Wimtrap"),
                      CDS = system.file("cds_example.bed", package = "Wimtrap"),
                      Intron = system.file("intron_example.bed", package = "Wimtrap"),
                      X3UTR = system.file("x3utr_example.bed", package = "Wimtrap")
                     )
imported_genomic_data.ex <- importGenomicData(biomart = FALSE,
                                              genomic_data = genomic_data.ex2,
                                              tss = system.file("tss_example.bed", package = "Wimtrap"),
                                              tts = system.file("tts.bed", package = "Wimtrap"))
                                              genomic_data = genomic_data.ex)

##With automatic download from Biomart of data related to gene structure
genomic_data.ex <- c(CE = system.file("conserved_elements_example.bed", package = "Wimtrap"),
                     DGF = system.file("DGF_example.bed", package = "Wimtrap"),
                     DHS = system.file("DHS_example.bed", package = "Wimtrap")
                     )
imported_genomic_data.ex <- importGenomicData(organism = "Arabidopsis thaliana",
                                              biomart = TRUE,
                                              genomic_data = genomic_data.ex)
}
